+++ eng/eng.trans.js
@@ -1,5 +1,5 @@
 (function(){
-  var Trans = {};
+  let Trans = {};
   if ( typeof module === "object" && module && typeof module.exports === "object" ) {
     Trans = require("../trans.js");
     module.exports = EngTrans;
@@ -8,7 +8,7 @@
     window.JsLingua.addService("Trans", "eng", EngTrans);
   }
 
-  var english = [
+  const english = [
     "A",
     "B",
     "C",
@@ -35,10 +35,8 @@
     "X",
     "Y",
     "Z"
-  ]
-
-  //https://en.wikipedia.org/wiki/Morse_code_for_non-Latin_alphabets
-  var enmorse = [
+  ],
+  enmorse = [//https://en.wikipedia.org/wiki/Morse_code_for_non-Latin_alphabets
     " .- ", // A
     " -... ", // B
     " -.-. ", // C
@@ -65,9 +63,8 @@
     " -..- ", // X
     " -.-- ", // Y
     " --.. " // Z
-  ];
-
-  var otherMourseBef = [
+  ],
+  otherMourseBef = [
     //numbers,
     "0",
     "1",
@@ -98,9 +95,8 @@
     "\"",
     "$",
     "@"
-  ];
-
-  var otherMourseAft = [
+  ],
+  otherMourseAft = [
     //numbers
     " ----- ",
     " .---- ",
@@ -133,7 +129,7 @@
     " .--.-. "
   ];
 
-  var otherMourseTrans = new Trans("otherMourse");
+  let otherMourseTrans = new Trans("otherMourse");
   Trans.newMethod.call(otherMourseTrans, "def", otherMourseBef, otherMourseAft);
 
   /**
@@ -160,7 +156,7 @@
    * @return {string}      processed text for morse transliteration
    */
   function morsePreTrans(text){
-    var result = text;
+    let result = text;
     //cleaning non supported codes
     result = result.replace(/[\^\#]/gi, "");
     result = result.replace(/\./gi, "#").replace(/\-/gi, "^");
@@ -176,7 +172,7 @@
    * @return {string}      filtered morse code
    */
   function morsePostTrans(text){
-    var result = text;
+    let result = text;
     result = otherMourseTrans.transliterate(result);
     result = result.replace(/ +/gi, " ");
     result = result.replace(/^ /gi, "");
@@ -195,7 +191,7 @@
    * @return {string}      processed morse code for untransliteration
    */
   function morsePreUntrans(text){
-    var result = text;
+    let result = text;
     //clean non morse characters
     result = result.replace(/[^ \.\-]/gi, "");
     result = result.replace(/[ ]{3,}/gi, " \t ");
@@ -213,7 +209,7 @@
    * @return {string}      filtered Arabic text
    */
   function morsePostUntrans(text){
-    var result = text;
+    let result = text;
     result = otherMourseTrans.untransliterate(result);
     result = result.replace(/\#/gi, ".").replace(/\^/gi, "-");
     result = result.replace(/ +/gi, "");