+++ jpn/jpn.lang.js
@@ -1,6 +1,6 @@
 (function(){
 
-  var Lang = {};
+  let Lang = {};
   if ( typeof module === "object" && module && typeof module.exports === "object" ) {
     Lang = require("../lang.js");
     module.exports = JpnLang;
@@ -28,18 +28,17 @@
   }
 
   JpnLang.prototype = Object.create(Lang.prototype);
-  var Me = JpnLang.prototype;
+  let Me = JpnLang.prototype;
   Me.constructor = JpnLang;
 
 
-  var lookup = {
+  const lookup = {
     0: "零", 1:"一", 2:"二", 3:"三", 4:"四", 5:"五",
     6:"六", 7:"七", 8:"八", 9:"九", 10:"十",
     100:"百", 1000:"千", 10000:"万",
     100000000:"億", 1000000000000:"兆"
-  }
-
-  var bigNbr = [
+  },
+  bigNbr = [
     1000000000000, 100000000,
     10000, 1000, 100, 10
   ]
@@ -62,7 +61,7 @@
     if (isNaN(num))
     return "";
 
-    var neg = false;
+    let neg = false;
     if(num < 0){
       neg = true;
       num = - num;
@@ -72,26 +71,23 @@
     return lookup[num];
 
     //search for the max index
-    var i = 0;
-    var max = 1;
-    var div;
+    let i = 0,
+    max = 1,
+    div;
     for (i = 0; i < bigNbr.length; i++){
       max = bigNbr[i]
       div = ~~(num/max);
       if (div > 0)
       break;
     }
-    var rem = ~~(num % max);
-    var result = "";
+    let rem = ~~(num % max),
+    result = "";
     if (div > 0)
-    if (div > 1 || max > 1000)
-    result += toJapaneseLetters(div);
+    if (div > 1 || max > 1000)  result += toJapaneseLetters(div);
     result += lookup[max];
-    if(rem > 0)
-    result += toJapaneseLetters(rem);
+    if(rem > 0) result += toJapaneseLetters(rem);
 
-    if(neg)
-    result = "マイナス" + result;
+    if(neg) result = "マイナス" + result;
 
     return result;