+++ tools.js
@@ -16,21 +16,24 @@
  * @param  {string} morse morse code written using ".", "-" and blacks " ".
  */
 function readMorse(morse){
-  var dotLength = 0.07;
-  var AudioContext = window.AudioContext || window.webkitAudioContext;
-  var context = new AudioContext();
-  var gain = context.createGain();
+  let dotLength = 0.07,
+  AudioContext = window.AudioContext || window.webkitAudioContext,
+  context = new AudioContext(),
+  gain = context.createGain(),
+  osc = context.createOscillator();
+
   gain.gain.value = 0;
-  var osc = context.createOscillator();
+
   osc.frequency.value = 600;
   osc.connect(gain);
   gain.connect(context.destination);
   osc.start(0);
-  var i;
-  var time0 = context.currentTime;
-  var time = time0;
+  let i,
+  time0 = context.currentTime,
+  time = time0;
+
   for (i =0; i< morse.length; i++){
-    var char = morse.charAt(i);
+    let char = morse.charAt(i);
     switch (char) {
       case ".":
         gain.gain.setValueAtTime(1.0, time);
@@ -62,7 +65,7 @@ function readMorse(morse){
  * @return {string}     skip string
  */
 function htmlEntities(str) {
-  var result = String(str).replace(/&/g, '&amp;').replace(/</g, '&lt;');
+  let result = String(str).replace(/&/g, '&amp;').replace(/</g, '&lt;');
   result = result.replace(/>/g, '&gt;').replace(/"/g, '&quot;').replace(/ /gi, "&nbsp;");
   return result;
 }